requirements.txt
APScheduler==3.9.1
asgiref==3.5.2
async-timeout==4.0.2
baidu-aip==4.16.7
certifi==2022.6.15
chardet==5.0.0
charset-normalizer==2.1.1
cloudscraper==1.2.60
defusedxml==0.7.1
deluge-client==1.9.0
Deprecated==1.2.13
diff-match-patch==20200713
Django==4.1
django-apscheduler==0.6.2
django-import-export==2.8.0
django-redis==5.2.0
django-simpleui==2022.7.29
et-xmlfile==1.1.0
idna==3.3
importlib-metadata==4.12.0
lxml==4.9.1
Markdown==3.4.1
MarkupPy==1.14
odfpy==1.4.1
OpenCC==1.1.4
openpyxl==3.0.10
packaging==21.3
pyparsing==3.0.9
pypushdeer==0.0.3
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qbittorrent-api==2022.8.36
redis==4.3.4
requests==2.28.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
transmission-rpc==3.3.2
typing_extensions==4.3.0
tzdata==2022.2
tzlocal==4.2
urllib3==1.26.11
wechat-push==1.0.1
wrapt==1.14.1
wxpusher==2.2.0
xlrd==2.0.1
xlwt==1.3.0
zipp==3.8.1
